A Capture is a Pattern and therefor should be an alias for Pattern.
Regarding lua-language-server it might be helpful to set
Lua.hover.expandAlias to false in the configuration file.
For example the tooltip for lpeg.Cc() will tell you then that
the type of the return value is Capture but it will be
accepted as a Pattern (for e.g. other function arguments) too.

https://github.com/Josef-Friedrich/LuaTeX_Lua-API/pull/4